backend/requirements.txt
# AI Knowledge Engine Backend Dependencies

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Machine Learning and Embeddings
sentence-transformers==2.2.2
torch==2.1.1
numpy==1.24.4
scikit-learn==1.3.2

# Vector Search
faiss-cpu==1.7.4

# Language Detection
fasttext==0.9.2
langdetect==1.0.9

# Topic Modeling
bertopic==0.16.0

# Data Processing
pandas==2.1.3
plotly==5.18.0

# Hugging Face Integration
huggingface_hub==0.20.0
datasets==2.16.0

# Text Processing
nltk==3.8.1
spacy==3.7.2

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # Used by FastAPI TestClient
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Date and time utilities
python-dateutil==2.8.2

# Async support
asyncio-mqtt==0.13.0
aiofiles==23.2.1

# JSON handling
orjson==3.9.10

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File handling
pillow==10.1.0

# Configuration
pyyaml==6.0.1
toml==0.10.2
